+++ admin/ce/components/authentication/authentication-modes.tsx
@@ -10,7 +10,7 @@ import {
 // components
 import { AuthenticationMethodCard } from "@/components/authentication";
 // helpers
-import { getBaseAuthenticationModes } from "@/helpers/authentication.helper";
+import { getBaseAuthenticationModes } from "@/lib/auth-helpers";
 // plane admin components
 import { UpgradeButton } from "@/plane-admin/components/common";
 // images
