+++ admin/core/components/admin-sidebar/sidebar-menu.tsx
@@ -4,11 +4,11 @@ import { observer } from "mobx-react";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import { Image, BrainCog, Cog, Lock, Mail } from "lucide-react";
+// plane internal packages
 import { Tooltip, WorkspaceIcon } from "@plane/ui";
+import { cn } from "@plane/utils";
 // hooks
-import { cn } from "@/helpers/common.helper";
 import { useTheme } from "@/hooks/store";
-// helpers
 
 const INSTANCE_ADMIN_LINKS = [
   {
