+++ admin/core/components/authentication/authentication-method-card.tsx
@@ -2,7 +2,7 @@
 
 import { FC } from "react";
 // helpers
-import { cn } from "helpers/common.helper";
+import { cn } from "@plane/utils";
 
 type Props = {
   name: string;
