+++ admin/core/components/authentication/google-config.tsx
@@ -5,12 +5,10 @@ import { observer } from "mobx-react";
 import Link from "next/link";
 // icons
 import { Settings2 } from "lucide-react";
-// types
+// plane internal packages
 import { TInstanceAuthenticationMethodKeys } from "@plane/types";
-// ui
 import { ToggleSwitch, getButtonStyling } from "@plane/ui";
-// helpers
-import { cn } from "@/helpers/common.helper";
+import { cn } from "@plane/utils";
 // hooks
 import { useInstance } from "@/hooks/store";
