+++ admin/core/services/auth.service.ts
@@ -1,5 +1,4 @@
-// helpers
-import { API_BASE_URL } from "@/helpers/common.helper";
+import { API_BASE_URL } from "@plane/constants";
 // services
 import { APIService } from "@/services/api.service";
