--- admin/helpers/index.ts
@@ -1,2 +1,0 @@
-export * from "./instance.helper";
-export * from "./user.helper";
