+++ packages/constants/src/index.ts
@@ -1,4 +1,8 @@
 export * from "./auth";
 export * from "./endpoints";
+export * from "./instance";
 export * from "./issue";
+export * from "./metadata";
+export * from "./swr";
+export * from "./user";
 export * from "./workspace";
