+++ packages/constants/src/metadata.ts
@@ -1,8 +1,11 @@
-export const SITE_NAME = "Plane | Simple, extensible, open-source project management tool.";
-export const SITE_TITLE = "Plane | Simple, extensible, open-source project management tool.";
+export const SITE_NAME =
+  "Plane | Simple, extensible, open-source project management tool.";
+export const SITE_TITLE =
+  "Plane | Simple, extensible, open-source project management tool.";
 export const SITE_DESCRIPTION =
   "Open-source project management tool to manage issues, sprints, and product roadmaps with peace of mind.";
 export const SITE_KEYWORDS =
   "software development, plan, ship, software, accelerate, code management, release management, project management, issue tracking, agile, scrum, kanban, collaboration";
 export const SITE_URL = "https://app.plane.so/";
-export const TWITTER_USER_NAME = "Plane | Simple, extensible, open-source project management tool.";
+export const TWITTER_USER_NAME =
+  "Plane | Simple, extensible, open-source project management tool.";
