+++ packages/utils/src/file.ts
@@ -1,5 +1,4 @@
-// helpers
-import { API_BASE_URL } from "@/helpers/common.helper";
+import { API_BASE_URL } from "@plane/constants";
 
 /**
  * @description combine the file path with the base URL
