+++ packages/utils/src/index.ts
@@ -1,4 +1,7 @@
+export * from "./auth";
 export * from "./color";
 export * from "./common";
 export * from "./emoji";
+export * from "./file";
 export * from "./string";
+export * from "./theme";
